+++ src/main/java/org/bukkit/material/Wood.java
@@ -51,7 +51,7 @@ public class Wood extends MaterialData {
     /**
      * Constructs a wood block of the given type and tree species.
      *
-     * @param type the type of wood block
+     * @param type    the type of wood block
      * @param species the species of the wood block
      */
     public Wood(final Material type, final TreeSpecies species) {
@@ -83,7 +83,7 @@ public class Wood extends MaterialData {
     /**
      * Correct the block type for certain species-type combinations.
      *
-     * @param type The desired type
+     * @param type    The desired type
      * @param species The required species
      * @return The actual type for this species given the desired type
      */
@@ -124,7 +124,7 @@ public class Wood extends MaterialData {
         switch (getItemType()) {
             case WOOD:
             case WOOD_DOUBLE_STEP:
-                return TreeSpecies.getByData((byte) getData());
+                return TreeSpecies.getByData(getData());
             case LOG:
             case LEAVES:
                 return TreeSpecies.getByData((byte) (getData() & 0x3));
